--- a/webrender/Cargo.toml
+++ b/webrender/Cargo.toml
@@ -11,6 +11,7 @@ default = ["freetype-lib"]
 freetype-lib = ["freetype/servo-freetype-sys"]
 profiler = ["thread_profiler/thread_profiler"]
 debugger = ["ws", "serde_json", "serde", "serde_derive"]
+query = []
 
 [dependencies]
 app_units = "0.5.6"