Fixes #1860

Feature is disabled by default, but building is CI-tested. Not sure how to make it enabled for, say, debug on desktop?..

This PR comes with an assumption that Gecko will know when to build with it (e.g. for developer builds) and when to not (on ARM).

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/servo/webrender/1861)
<!-- Reviewable:end -->