@glennw 2017-10-13
CI failures:

```
error: associated constants are experimental (see issue #29646)

   --> src/device.rs:557:5

    |

557 |     const MAX_TIMERS_PER_FRAME: usize = 256;

    |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

error: associated constants are experimental (see issue #29646)

   --> src/device.rs:560:5

    |

560 |     const MAX_SAMPLERS_PER_FRAME: usize = 0;

    |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

error: aborting due to previous error(s)
```