@bors-servo 2017-10-17
:sunny: Test successful - [status-appveyor](https://ci.appveyor.com/project/servo/webrender/build/1.0.5850), [status-travis](https://travis-ci.org/servo/webrender/builds/288842658?utm_source=github_status&utm_medium=notification)
Approved by: glennw
Pushing 5e30fb4168a22440d91066c246eb16e35eb84e91 to master...