@glennw 2017-10-17
@kvark Did some performance tests of this change.

With query timers enabled:
 Wikipedia compositor CPU time 2.0ms
 HN compositor CPU time 2.4ms

Without query timers enabled:
 Wikipedia compositor CPU time 1.3ms
 HN compositor CPU time 1.8ms

That's a nice, free performance win!