@staktrace 2017-10-17
@kvark for some reason this PR is causing cbindgen to fail when parsing the webrender crate. It looks more like a cbindgen problem than a problem with your patch but just a heads-up anyway. Details at https://bugzilla.mozilla.org/show_bug.cgi?id=1408461#c8 onwards.