@staktrace 2017-10-17
Filed #1887 with a possible fix